maa-core: static link fastdeploy_ppocr and opencv when needed
OpenCV is a very large dependency, and we only need a small part of it,
so we build our own opencv if user does not want to install opencv by homebrew.
To avoid the conflict with the opencv installed by homebrew,
we build our own opencv and fastdeploy_ppocr and link them statically.

Closes #125.

Signed-off-by: Loong <40141251+wangl-cc@users.noreply.github.com>